Split behavior association replay tests
